% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{form1}
\alias{form1}
\title{Form 1 from a real credentialing dataset}
\format{
A data frame with 1,636 rows and 32 columns.
\describe{
\item{EID}{unique person ID variable}
\item{cent_id}{unique test center ID variable}
\item{iraw.A1}{dichotomous item score for item A1}
\item{iraw.A2}{dichotomous item score for item A2}
\item{iraw.A3}{dichotomous item score for item A3}
\item{iraw.A4}{dichotomous item score for item A4}
\item{iraw.A5}{dichotomous item score for item A5}
\item{iraw.A6}{dichotomous item score for item A6}
\item{iraw.A7}{dichotomous item score for item A7}
\item{iraw.A8}{dichotomous item score for item A8}
\item{iraw.A9}{dichotomous item score for item A9}
\item{iraw.A10}{dichotomous item score for item A10}
\item{iraw.A11}{dichotomous item score for item A11}
\item{iraw.A12}{dichotomous item score for item A12}
\item{iraw.A13}{dichotomous item score for item A13}
\item{iraw.A14}{dichotomous item score for item A14}
\item{iraw.A15}{dichotomous item score for item A15}
\item{iraw.A16}{dichotomous item score for item A16}
\item{iraw.A17}{dichotomous item score for item A17}
\item{iraw.A18}{dichotomous item score for item A18}
\item{iraw.A19}{dichotomous item score for item A19}
\item{iraw.A20}{dichotomous item score for item A20}
\item{iraw.A21}{dichotomous item score for item A21}
\item{iraw.A22}{dichotomous item score for item A22}
\item{iraw.A23}{dichotomous item score for item A23}
\item{iraw.A24}{dichotomous item score for item A24}
\item{iraw.A25}{dichotomous item score for item A25}
\item{iraw.A26}{dichotomous item score for item A26}
\item{iraw.A27}{dichotomous item score for item A27}
\item{iraw.A28}{dichotomous item score for item A28}
\item{iraw.A29}{dichotomous item score for item A29}
\item{iraw.A30}{dichotomous item score for item A30}
}
}
\source{
Cizek, G. J., & Wollack, J. A. (Eds.). (2017). Handbook of quantitative methods
for detecting cheating on tests. New York, NY: Routledge.
}
\usage{
form1
}
\description{
This is a subset of Form 1 from a real credentialing dataset provided by Cizek and Wollack (2017).
It has 1,636 test takers and their response to 30 items. The dataset also includes the
unique individual IDs in the first column and the unique center IDs in the second column.
Column 3 to Column 32 includes the dichotomous item responses to 30 items.
}
\keyword{datasets}
