\name{form2}
\alias{form2}
\docType{data}
\title{Form 2 of a real credentialing dataset}
\description{

This is the Form 2 of a real credentialing dataset provided by Cizek and Wollack (2017). It has 1,607 test takers and their response to 170 items. The dataset also includes the unique individual IDs in the first column and the unique center IDs in the second column. Column 3 to Column 172 includes the nominal item responses to 170 items.

}
\usage{data(form2)}
\format{
  A data frame with 1,607 rows and 172 columns.
}
\references{

Cizek, G. J., & Wollack, J. A. (Eds.). (2017). Handbook of quantitative methods for detecting cheating on tests. New York, NY: Routledge.

}

