\name{ipar}
\alias{ipar}
\docType{data}
\title{
The Nominal Response IRT Model Item Parameters
}
\description{
The Nominal Response Model (NRM) item parameters are taken from Wollack's study (1996) and used to generate \code{\link{simulated.data}}.
}
\usage{data(ipar)}
\format{
  A data frame with 40 rows and 10 columns.
  \describe{
    \item{\code{Slope.A}}{NRM slope parameter for response alternative A}
    \item{\code{Slope.B}}{NRM slope parameter for response alternative B}
    \item{\code{Slope.C}}{NRM slope parameter for response alternative C}
    \item{\code{Slope.D}}{NRM slope parameter for response alternative D}
    \item{\code{Slope.E}}{NRM slope parameter for response alternative E}
    \item{\code{Intercept.A}}{NRM intercept parameter for response alternative A}
    \item{\code{Intercept.B}}{NRM intercept parameter for response alternative B}
    \item{\code{Intercept.C}}{NRM intercept parameter for response alternative C}
    \item{\code{Intercept.D}}{NRM intercept parameter for response alternative D}
    \item{\code{Intercept.E}}{NRM intercept parameter for response alternative E}
  }
}
\references{
Wollack, J.A.(1996). Detection of answer copying using item response theory. \emph{Dissertation Abstracts International, 57/05}, 2015.\cr

Bock (1972). Estimating item parameters and latent ability when responses are scored in two or more nominal categories. \emph{Psychometrika, 37}, 29-51 .
\cr
\cr
\cr
\cr
\cr
}

