% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{negocios}
\alias{negocios}
\title{Exam responses from a mastery test for a graduate class}
\format{
A data frame with 500 rows and 61 columns.
\describe{
\item{ID}{unique person ID variable}
\item{P1}{nominal item score for P1}
\item{P2}{nominal item score for P2}
\item{P3}{nominal item score for P3}
\item{P4}{nominal item score for P4}
\item{P5}{nominal item score for P5}
\item{P6}{nominal item score for P6}
\item{P7}{nominal item score for P7}
\item{P8}{nominal item score for P8}
\item{P9}{nominal item score for P9}
\item{P10}{nominal item score for P10}
\item{P11}{nominal item score for P11}
\item{P12}{nominal item score for P12}
\item{P13}{nominal item score for P13}
\item{P14}{nominal item score for P14}
\item{P15}{nominal item score for P15}
\item{P16}{nominal item score for P16}
\item{P17}{nominal item score for P17}
\item{P18}{nominal item score for P18}
\item{P19}{nominal item score for P19}
\item{P20}{nominal item score for P20}
\item{P21}{nominal item score for P21}
\item{P22}{nominal item score for P22}
\item{P23}{nominal item score for P23}
\item{P24}{nominal item score for P24}
\item{P25}{nominal item score for P25}
\item{P26}{nominal item score for P26}
\item{P27}{nominal item score for P27}
\item{P28}{nominal item score for P28}
\item{P29}{nominal item score for P29}
\item{P30}{nominal item score for P30}
\item{P31}{nominal item score for P31}
\item{P32}{nominal item score for P32}
\item{P33}{nominal item score for P33}
\item{P34}{nominal item score for P34}
\item{P35}{nominal item score for P35}
\item{P36}{nominal item score for P36}
\item{P37}{nominal item score for P37}
\item{P38}{nominal item score for P38}
\item{P39}{nominal item score for P39}
\item{P40}{nominal item score for P40}
\item{P41}{nominal item score for P41}
\item{P42}{nominal item score for P42}
\item{P43}{nominal item score for P43}
\item{P44}{nominal item score for P44}
\item{P45}{nominal item score for P45}
\item{P46}{nominal item score for P46}
\item{P47}{nominal item score for P47}
\item{P48}{nominal item score for P48}
\item{P49}{nominal item score for P49}
\item{P50}{nominal item score for P50}
\item{P51}{nominal item score for P51}
\item{P52}{nominal item score for P52}
\item{P53}{nominal item score for P53}
\item{P54}{nominal item score for P54}
\item{P55}{nominal item score for P55}
\item{P56}{nominal item score for P56}
\item{P57}{nominal item score for P57}
\item{P58}{nominal item score for P58}
\item{P59}{nominal item score for P59}
\item{P60}{nominal item score for P60}
}
}
\source{
\url{http://lertap5.com/HTMLHelp/Lrtp59HTML/index.html?negocios.htm}
}
\usage{
negocios
}
\description{
This data set involves exam responses from a postgraduate class in business
studies. Cheating was suspected by one of the exam proctors. When confronted
with evidence, two students confessed that they had cheated by sharing answers
during the administration of the test.
}
\details{
Key response vector is:

3343233133 1221411113 4224232323 4323341411 4244232422 3232113313
}
\keyword{datasets}
