\name{simulated.data}
\alias{simulated.data}
\docType{data}
\title{
Simulated Data based on Nominal Response IRT Model (Bock, 1972)
}
\description{
Data with 40 items is simulated based on the Nominal Response IRT Model (Bock, 1972). Each variable (column) represents an item and is a "character" variable. This is necessary to be able to run the \code{\link{CopyDetect2}} function. The items are assumed to have five response alternatives: A, B, C, D, and E. 

The item parameters used to generate data are taken from Wollack's study (1996). 
}
\usage{data(simulated.data)}
\format{
  A data frame with 500 hypothetical simulees.
}
\references{
Wollack, J.A.(1996). Detection of answer copying using item response theory. \emph{Dissertation Abstracts International, 57/05}, 2015.\cr

Bock (1972). Estimating item parameters and latent ability when responses are scored in two or more nominal categories. \emph{Psychometrika, 37}, 29-51 .
\cr
\cr
\cr
\cr
\cr
}

