% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{form2}
\alias{form2}
\title{Form 2 of a real credentialing dataset}
\format{
A data frame with 1,607 rows and 32 columns.
\describe{
\item{EID}{unique person ID variable}
\item{cent_id}{unique test center ID variable}
\item{iresp.B1}{nominal item score for item B1}
\item{iresp.B2}{nominal item score for item B2}
\item{iresp.B3}{nominal item score for item B3}
\item{iresp.B4}{nominal item score for item B4}
\item{iresp.B5}{nominal item score for item B5}
\item{iresp.B6}{nominal item score for item B6}
\item{iresp.B7}{nominal item score for item B7}
\item{iresp.B8}{nominal item score for item B8}
\item{iresp.B9}{nominal item score for item B9}
\item{iresp.B10}{nominal item score for item B10}
\item{iresp.B11}{nominal item score for item B11}
\item{iresp.B12}{nominal item score for item B12}
\item{iresp.B13}{nominal item score for item B13}
\item{iresp.B14}{nominal item score for item B14}
\item{iresp.B15}{nominal item score for item B15}
\item{iresp.B16}{nominal item score for item B16}
\item{iresp.B17}{nominal item score for item B17}
\item{iresp.B18}{nominal item score for item B18}
\item{iresp.B19}{nominal item score for item B19}
\item{iresp.B20}{nominal item score for item B20}
\item{iresp.B21}{nominal item score for item B21}
\item{iresp.B22}{nominal item score for item B22}
\item{iresp.B23}{nominal item score for item B23}
\item{iresp.B24}{nominal item score for item B24}
\item{iresp.B25}{nominal item score for item B25}
\item{iresp.B26}{nominal item score for item B26}
\item{iresp.B27}{nominal item score for item B27}
\item{iresp.B28}{nominal item score for item B28}
\item{iresp.B29}{nominal item score for item B29}
\item{iresp.B30}{nominal item score for item B30}
}
}
\source{
Cizek, G. J., & Wollack, J. A. (Eds.). (2017). Handbook of quantitative methods
for detecting cheating on tests. New York, NY: Routledge.
}
\usage{
form2
}
\description{
This is a subset of Form 2 from a real credentialing dataset provided by Cizek and Wollack (2017).
It has 1,607 test takers and their response to 30 items. The dataset also includes the
unique individual IDs in the first column and the unique center IDs in the second column.
Column 3 to Column 32 includes the nominal item responses to 30 items.
}
\keyword{datasets}
